requirements.txt
akismet==1.1
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-admin-rangefilter==0.6.2
django-allauth==0.42.0
django-anymail==7.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==2.0.0
django-bootstrap4==2.2.0
django-contact-form==1.8.1
django-countries==6.1.2
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.5
phonenumbers==8.12.6
Pillow==7.2.0
pip==20.1.1
psycopg2==2.8.5
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
setuptools==47.3.1.post20200616
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
wheel==0.34.2
whitenoise==5.1.0
wincertstore==0.2
zipp==3.1.0
